requirements.txt
torch>=2.1.0
ultralytics>=8.0.0
pillow
pdf2image
PyMuPDF
tqdm
